+++ tests/test_daemon_modules_http.py
@@ -1,13 +1,14 @@
+import pytest
 import os
-import sys
 import time
 import json
 import shutil
 import requests
 import zipfile
-import unittest
-sys.path.append("../")
-from dorc.util import make_test_daemon
+
+from util import assertIn
+from dorc.daemon.models import SessionMethodResponseModel
+smod = SessionMethodResponseModel
 
 
 mod_string = """
@@ -113,195 +114,148 @@ def hack_param():
 """
 
 
-class DaemonModulesTest(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        cls.data_dir = ".test_dir"
-        if os.path.exists(cls.data_dir):
-            shutil.rmtree(cls.data_dir)
-        if not os.path.exists(cls.data_dir):
-            os.mkdir(cls.data_dir)
-        cls.port = 23232
-        cls.hostname = "127.0.0.1"
-        cls.daemon, cls.cookies = make_test_daemon(cls.hostname, cls.port, ".test_dir",
-                                                   name="test@local", get_cookies=True)
-        cls.host = "http://" + ":".join([cls.hostname, str(cls.port) + "/"])
-        time.sleep(.5)
-        with open("_setup.py", "rb") as f:
-            response = requests.request("POST", cls.host + "create_session",
-                                        files={"file": f},
-                                        data={"name": json.dumps("meh_session")},
-                                        cookies=cls.cookies)
-            print(response.content)
-        time.sleep(1)
-        with open("_setup.py", "rb") as f:
-            response = requests.request("POST", cls.host + "create_session",
-                                        files={"file": f},
-                                        data={"name": json.dumps("bleh_session")},
-                                        cookies=cls.cookies)
-            print(response.content)
-        if cls.daemon._fwd_ports_thread is not None:
-            cls.daemon._fwd_ports_thread.kill()
-        time.sleep(1)
-        print(cls.daemon.daemon_name)
-        print("Initialized daemon and created sessions")
+def test_list_modules(daemon_and_cookies, json_config):
+    daemon, cookies = daemon_and_cookies
+    time.sleep(1)
+    host = f"http://{daemon.hostname}:{daemon.port}/"
+    response = requests.request("GET", host + "list_global_modules", cookies=cookies)
+    assert isinstance(json.loads(response.content), list)
+    assert isinstance(json.loads(response.content)[1], dict)
+    mods = json.loads(response.content)
+    expected = {"autoloads": ["torch", "ModelStep", "sys",
+                              "ClassificationStep",
+                              "accuracy", "CheckFunc",
+                              "CheckGreater",
+                              "CheckGreaterName",
+                              "CheckLesserName",
+                              "CheckAccuracy"]}
+    for k in mods[1]:
+        assert k in expected
+        assert set(mods[1][k]) == set(expected[k])
 
-    def test_list_modules(self):
-        time.sleep(1)
-        response = requests.request("GET", self.host + "list_global_modules", cookies=self.cookies)
-        self.assertIsInstance(json.loads(response.content), list)
-        self.assertIsInstance(json.loads(response.content)[1], dict)
-        mods = json.loads(response.content)
-        # NOTE: Not sure why this fails
-        # self.assertEqual(mods, {"autoloads": ["torch", "ModelStep",
-        #                                       "ClassificationTrainStep",
-        #                                       "ClassificationTestStep",
-        #                                       "accuracy", "CheckFunc",
-        #                                       "CheckGreater",
-        #                                       "CheckGreaterName",
-        #                                       "CheckLesserName",
-        #                                       "CheckAccuracy"]})
 
-    def test_add_delete_global_module(self):
-        with open("./._test_py_file.py", "w") as f:
-            f.write(mod_string)
-        with open("./._test_py_file.py", "rb") as f:
-            response = requests.request("POST", self.host + "add_global_module",
-                                        files={"file": f},
-                                        data={"name": json.dumps("test_module_a")},
-                                        cookies=self.cookies)
-        self.assertIsInstance(json.loads(response.content), list)
-        self.assertIsInstance(json.loads(response.content)[1], dict)
-        status, result = json.loads(response.content)
-        self.assertTrue(status)
-        task_id = result["task_id"]
-        time.sleep(1)
-        response = requests.request("GET", self.host + f"check_task?task_id={task_id}",
-                                    cookies=self.cookies)
-        status, result = json.loads(response.content)
-        self.assertTrue(result["result"])
-        self.assertIn("_module_test_module_a", self.daemon._modules)
-        with open("./._test_py_file.py", "rb") as f:
-            response = requests.request("POST", self.host + "add_global_module",
-                                        files={"file": f},
-                                        data={"name": json.dumps("test_module_b")},
-                                        cookies=self.cookies)
-        status, result = json.loads(response.content)
-        self.assertTrue(status)
-        task_id = result["task_id"]
-        time.sleep(1)
-        response = requests.request("GET", self.host + f"check_task?task_id={task_id}",
-                                    cookies=self.cookies)
-        status, result = json.loads(response.content)
-        self.assertTrue(result["result"])
-        self.assertIn("_module_test_module_b", self.daemon._modules)
-        requests.request("POST", self.host + "delete_global_module",
-                         data={"name": "test_module_a"},
-                         cookies=self.cookies)
-        time.sleep(.3)
-        self.assertTrue(result["result"])
-        self.assertNotIn("_module_test_module_a", self.daemon._modules)
+def test_add_delete_global_module(daemon_and_cookies, json_config):
+    daemon, cookies = daemon_and_cookies
+    host = f"http://{daemon.hostname}:{daemon.port}/"
+    with open("./._test_py_file.py", "w") as f:
+        f.write(mod_string)
+    with open("./._test_py_file.py", "rb") as f:
+        response = requests.request("POST", host + "add_global_module",
+                                    files={"file": f},
+                                    data={"name": json.dumps("test_module_a")},
+                                    cookies=cookies)
+    result = smod.parse_obj(json.loads(response.content))
+    assert result.result
+    task_id = result.task_id
+    time.sleep(1)
+    response = requests.request("GET", host + f"check_task?task_id={task_id}",
+                                cookies=cookies)
+    result = smod.parse_obj(json.loads(response.content))
+    assert result.result
+    assertIn("_module_test_module_a", daemon._modules)
+    with open("./._test_py_file.py", "rb") as f:
+        response = requests.request("POST", host + "add_global_module",
+                                    files={"file": f},
+                                    data={"name": json.dumps("test_module_b")},
+                                    cookies=cookies)
+    result = smod.parse_obj(json.loads(response.content))
+    assert result.result
+    task_id = result.task_id
+    time.sleep(1)
+    response = requests.request("GET", host + f"check_task?task_id={task_id}",
+                                cookies=cookies)
+    result = smod.parse_obj(json.loads(response.content))
+    assert result.result
+    assertIn("_module_test_module_b", daemon._modules)
+    response = requests.request("POST", host + "delete_global_module",
+                                data={"name": "test_module_a"},
+                                cookies=cookies)
+    time.sleep(.3)
+    result = response.content
+    print(result)
+    assert "_module_test_module_a" not in daemon._modules
 
-    def test_add_delete_dataset(self):
-        py_string = """
+
+def test_add_delete_dataset(daemon_and_cookies, json_config):
+    py_string = """
 class MnistDataset:
     def __init__(self, pt_file):
         import torch
-        self.data = torch.load(pt_file)
+        data = torch.load(pt_file)
 
-    def __len__(self):
-        return len(self.data[0])
+    def __len__(daemon_and_cookies, json_config):
+        return len(data[0])
 
     def __getitem__(self, idx):
-        return self.data[0][idx], self.data[1][idx]
+        return data[0][idx], data[1][idx]
 
 import os
 file_dir = os.path.abspath(os.path.dirname(__file__))
 dataset = MnistDataset(os.path.join(file_dir, "training.pt"))
 """
-        if not os.path.exists("./._test_data_dir"):
-            os.mkdir("./._test_data_dir")
-        t_dir = "./._test_data_dir"
-        with open(os.path.join(t_dir, "__init__.py"), "w") as f:
-            f.write(py_string)
-        shutil.copy(".data/MNIST/processed/training.pt", t_dir)
-        zf = zipfile.ZipFile(".test.zip", "w", zipfile.ZIP_DEFLATED)
-        for f in os.listdir(t_dir):
-            zf.write(os.path.join(t_dir, f), arcname=f)
-        zf.close()
-        with open("./.test.zip", "rb") as f:
-            response = requests.request("POST", self.host + "upload_dataset",
-                                        files={"file": f},
-                                        data={"name": "MNIST",
-                                              "description": "MNIST Image dataset",
-                                              "type": "image"},
-                                        cookies=self.cookies)
-        self.assertIsInstance(json.loads(response.content), list)
-        self.assertIsInstance(json.loads(response.content)[1], dict)
-        status, result = json.loads(response.content)
-        self.assertTrue(status)
-        task_id = result["task_id"]
-        time.sleep(1)
-        response = requests.request("GET", self.host + f"check_task?task_id={task_id}",
-                                    cookies=self.cookies)
-        status, result = json.loads(response.content)
-        self.assertTrue(status)
-        self.assertTrue(result["result"])
-        self.assertIn("_dataset_MNIST", self.daemon._datasets)
-        py_string_2 = """
+    daemon, cookies = daemon_and_cookies
+    host = f"http://{daemon.hostname}:{daemon.port}/"
+    if not os.path.exists("./._test_data_dir"):
+        os.mkdir("./._test_data_dir")
+    t_dir = "./._test_data_dir"
+    with open(os.path.join(t_dir, "__init__.py"), "w") as f:
+        f.write(py_string)
+    shutil.copy(".data/MNIST/processed/training.pt", t_dir)
+    zf = zipfile.ZipFile(".test.zip", "w", zipfile.ZIP_DEFLATED)
+    for f in os.listdir(t_dir):
+        zf.write(os.path.join(t_dir, f), arcname=f)
+    zf.close()
+    with open("./.test.zip", "rb") as f:
+        response = requests.request("POST", host + "upload_dataset",
+                                    files={"file": f},
+                                    data={"name": "MNIST",
+                                          "description": "MNIST Image dataset",
+                                          "type": "image"},
+                                    cookies=cookies)
+    result = smod.parse_obj(json.loads(response.content))
+    assert result.result
+    task_id = result.task_id
+    time.sleep(1)
+    response = requests.request("GET", host + f"check_task?task_id={task_id}",
+                                cookies=cookies)
+    result = smod.parse_obj(json.loads(response.content))
+    assert result.result
+    assertIn("_dataset_MNIST", daemon._datasets)
+    py_string_2 = """
 class MnistDataset:
     def __init__(self, pt_file):
         import torch
-        self.data = torch.load(pt_file)
+        data = torch.load(pt_file)
 
-    def __len__(self):
-        return len(self.data[0])
+    def __len__(daemon_and_cookies, json_config):
+        return len(data[0])
 
     def __getitem__(self, idx):
-        return self.data[0][idx], self.data[1][idx]
+        return data[0][idx], data[1][idx]
 
 dataset = MnistDataset("/home/joe/projects/trainer/.data/MNIST/processed/training.pt")
 """
-        with open("./.data_file.py", "w") as f:
-            f.write(py_string_2)
-        with open("./.data_file.py", "rb") as f:
-            response = requests.request("POST", self.host + "upload_dataset",
-                                        files={"file": f},
-                                        data={"name": "MNIST_2",
-                                              "description": "MNIST Image dataset",
-                                              "type": "image"},
-                                        cookies=self.cookies)
-        self.assertIsInstance(json.loads(response.content), list)
-        self.assertIsInstance(json.loads(response.content)[1], dict)
-        status, result = json.loads(response.content)
-        self.assertTrue(status)
-        task_id = result["task_id"]
-        time.sleep(1)
-        response = requests.request("GET", self.host + f"check_task?task_id={task_id}",
-                                    cookies=self.cookies)
-        status, result = json.loads(response.content)
-        self.assertTrue(status)
-        self.assertTrue(result["result"])
-        self.assertIn("_dataset_MNIST_2", self.daemon._datasets)
-        requests.request("POST", self.host + "delete_dataset",
-                         data={"name": "MNIST_2"},
-                         cookies=self.cookies)
-        time.sleep(.3)
-        self.assertTrue(result["result"])
-        self.assertNotIn("_dataset_MNIST_2", self.daemon._datasets)
-
-    @classmethod
-    def shutdown_daemon(cls, host):
-        response = requests.request("GET", host + "_shutdown",
-                                    cookies=cls.cookies, timeout=2)
-        return response
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.shutdown_daemon(cls.host)
-        if os.path.exists(cls.data_dir):
-            shutil.rmtree(cls.data_dir)
-
-
-if __name__ == '__main__':
-    unittest.main()
+    with open("./.data_file.py", "w") as f:
+        f.write(py_string_2)
+    with open("./.data_file.py", "rb") as f:
+        response = requests.request("POST", host + "upload_dataset",
+                                    files={"file": f},
+                                    data={"name": "MNIST_2",
+                                          "description": "MNIST Image dataset",
+                                          "type": "image"},
+                                    cookies=cookies)
+    result = smod.parse_obj(json.loads(response.content))
+    assert result.result
+    task_id = result.task_id
+    time.sleep(1)
+    response = requests.request("GET", host + f"check_task?task_id={task_id}",
+                                cookies=cookies)
+    result = smod.parse_obj(json.loads(response.content))
+    assert result.result
+    assertIn("_dataset_MNIST_2", daemon._datasets)
+    response = requests.request("POST", host + "delete_dataset",
+                                data={"name": "MNIST_2"},
+                                cookies=cookies)
+    time.sleep(.3)
+    print(response.content)
+    assert "_dataset_MNIST_2" not in daemon._datasets