+++ tests/util.py
@@ -1,14 +1,30 @@
 import pytest
 import sys
-import shutil
-import datetime
-import os
+import copy
 import torch
 sys.path.append("../")
 from dorc.trainer import Trainer
 from dorc.trainer.model import Model
 from dorc.autoloads import ClassificationStep
 
+
+def assertIn(a, b):
+    assert a in b
+
+
+def terminate_live_sessions(daemon):
+    for s in daemon._sessions.values():
+        for s_name in s["sessions"]:
+            if "process" in s["sessions"][s_name]:
+                s["sessions"][s_name]["process"].terminate()
+                print(f'Terminated {s["sessions"][s_name]["process"]}')
+
+
+def _create_session(daemon, config, load=False):
+    data = {"name": "test_session", "config": copy.deepcopy(config), "load": load}
+    daemon.create_session(0, data)
+    result = daemon._check_result(0)
+    return result
 
 
 def make_daemon():